Extract cache key generation to separate file
