+++ api/app.py
@@ -4,6 +4,7 @@ from flask_caching import Cache
 from config import CACHE_CONFIG
 from datetime import datetime, timedelta
 
+from utils import calculate_date_cache_key
 from tweets import get_date_range, group_by_day, get_all_calculations, get_longest_gap, get_current_gap
 from dmi_tcat import fetch_tweets_for_date
 
@@ -13,15 +14,6 @@ CORS(app)
 app.config.from_mapping(CACHE_CONFIG)
 cache = Cache(app)
 
-def calculate_date_cache_key(date):
-  app.logger.debug(f'Calculating cache key for {date}')
-  if datetime.now().date() == datetime.strptime(date, '%Y-%m-%d').date():
-    period = int(int(datetime.strftime(datetime.now(), '%M')) / 15)
-    calculated_cache_key = datetime.strftime(datetime.now(), f'%Y-%m-%d %H {period}')
-  else:
-    calculated_cache_key = date
-  app.logger.debug(f'CALCULATED CACHE KEY: {calculated_cache_key}')
-  return calculated_cache_key
 
 @app.route('/<date>', methods=['GET'])
 @cache.cached(make_cache_key=calculate_date_cache_key)
@@ -31,11 +23,8 @@ def index(date):
 
   return jsonify(tweets=tweets, calculations=calculations)
 
-def get_summary_cache_key():
-  return datetime.now().strftime('%Y-%m-%d')
-
 @app.route('/summary', methods=['GET'])
-@cache.cached(make_cache_key=get_summary_cache_key)
+@cache.cached(make_cache_key=calculate_date_cache_key)
 def summary():
   start, end = get_date_range()
   tweets = fetch_tweets_for_date(start, end)
